+++ packages/durably/tests/browser/event-ordering.test.ts
@@ -1,0 +1,4 @@
+import { createBrowserDialect } from '../helpers/browser-dialect'
+import { createEventOrderingTests } from '../shared/event-ordering.shared'
+
+createEventOrderingTests(createBrowserDialect)