+++ packages/durably/tests/node/event-ordering.postgres.test.ts
@@ -1,0 +1,4 @@
+import { usePostgresSchemaPerTest } from '../helpers/postgres-dialect'
+import { createEventOrderingTests } from '../shared/event-ordering.shared'
+
+createEventOrderingTests(usePostgresSchemaPerTest())